# Week12/Day_2/2_app_builder/backend/api/endpoints/users.py
from fastapi import APIRouter, Depends, HTTPException, Path
from typing import List
from backend.models.schemas import Dashboard, Notification

router = APIRouter()

@router.get("/{userId}/dashboard", response_model=Dashboard, summary="View task progress dashboard")
async def view_dashboard(userId: str = Path(...)):
    # Implement actual logic
    pass

@router.get("/{userId}/notifications", response_model=List[Notification], summary="Retrieve deadline notifications")
async def get_notifications(userId: str = Path(...)):
    # Implement actual logic
    pass

@router.get("/{userId}/tasks", summary="Retrieve all tasks for a user")
async def get_tasks(userId: str = Path(...), status: str = None):
    # Implement actual logic
    pass

@router.post("/{userId}/tasks", summary="Create a new task")
async def create_task(userId: str, task_data: dict):
    # Implement actual logic
    pass

@router.get("/{userId}/reports", summary="Generate productivity report")
async def generate_report(userId: str = Path(...)):
    # Implement actual logic
    pass